Maintenance update (sort of)
* Cleaned part of the code, a little bit
* Added a Macros folder with little Macro headers that could be useful
* Set the Release mode to treat compilation warnings as errors

+++ src/SymphonyDemo/ComplexRenderer.cpp
@@ -159,7 +159,7 @@ void ComplexRenderer::RenderCamera(Camera* cam, const std::vector<OrderableObjec
         /* Update the shader with light data if it uses lights */
         /* Update the shader with light data only if object is within the light's reach (radius) */
         /* Update the shader with light data only if light gameobject is enabled */
-        int activeLights = 0;
+        size_t activeLights = 0;
         for (; activeLights < lights.size(); ++activeLights)
         {
             if (!lights[activeLights]->enabled) continue;
