+++ src/SymphonyDemo/SimpleRenderer.cpp
@@ -85,7 +85,7 @@ void SimpleRenderer::RenderCamera(Camera* cam, const std::vector<OrderableObject
         /* Update the shader with light data if it uses lights */
         /* Update the shader with light data only if object is within the light's reach (radius) */
         /* Update the shader with light data only if light gameobject is enabled */
-        int activeLights = 0;
+        size_t activeLights = 0;
         for (; activeLights < lights.size(); ++activeLights)
         {
             if (!lights[activeLights]->enabled) continue;
