+++ src/SymphonyEngine/Engine/SymphonyEngine.cpp
@@ -293,7 +293,7 @@ namespace Symphony
                 texture,
                 glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
                 glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-                face->glyph->advance.x,
+                (unsigned int)face->glyph->advance.x,
                 face->glyph->bitmap.pitch
             };
